fix(v11-batch37): 修复 custom_order_process_service 类型不匹配（i32→i64）
process_node.id 和 process_log.process_node_id 均为 i64，
logs_map 的 HashMap 键类型应声明为 i64 而非 i32。

CI 构建报错：error[E0308]: mismatched types, expected i32, found i64

+++ backend/src/services/custom_order_process_service.rs
@@ -211,7 +211,7 @@ impl CustomOrderProcessService {
                 .all(&*self.db)
                 .await?
         };
-        let mut logs_map: std::collections::HashMap<i32, Vec<process_log::Model>> =
+        let mut logs_map: std::collections::HashMap<i64, Vec<process_log::Model>> =
             std::collections::HashMap::new();
         for log in all_logs {
             logs_map.entry(log.process_node_id).or_default().push(log);
